Swing Type Module
Based on past pitches, classifies as right, left, middle, or spray hitter.

One thing to note, is we decided to sometime classify one hit as two types, which could mess up the averages (since total (bottom part of denominator to get percentages) is based on number of hits right now). We could increment total to count as total number swing classifications, this would get percentages adding up to 100, not sure if there is a downside to this.